# api/swagger.json
{
   "swagger":"2.0",
   "info":{
      "title":"Simple Weather API",
      "description":"API for getting the current weather information.",
      "version":"1.0"
   },
   "produces":[
      "application/json"
   ],
   "consumes":[
      "application/json"
   ],
   "host":"localhost:3000",
   "basePath":"/service-provider/api",
   "paths":{
      "/registation":{
         "post":{
            "x-swagger-router-controller":"AuthenticationController",
            "operationId":"verifyMobile",
            "tags":[
               "/registation"
            ],
            "summary":"Verify Service Provider Mobile Number",
            "description":"Verify Mobile Number.",
            "parameters":[
               {
                  "name":"mobileNumber",
                  "in":"body",
                  "description":"mobile number to verify.",
                  "schema":{
                     "required":[
                        "mobileNumber"
                     ],
                     "properties":{
                        "mobileNumber":{
                           "type":"string"
                        }
                     }
                  }
               }
            ],
            "responses":{
               "201":{
                  "description":"Success."
               },
               "400":{
                  "description":"ERROR."
               }
            }
         }
      },
      "/verfication/verify":{
         "post":{
            "x-swagger-router-controller":"AuthenticationController",
            "operationId":"verified",
            "tags":[
               "/registation"
            ],
            "summary":"Verfy  submition",
            "description":"Verfy .",
            "parameters":[
               {
                  "name":"verify",
                  "in":"body",
                  "description":"verficationCode to verify.",
                  "schema":{
                     "required":[
                        "verify"
                     ],
                     "properties":{
                        "verify":{
                           "type":"object",
                           "properties":{
                              "mobileNumber":{
                                 "type":"string"
                              },
                              "verifyCode":{
                                 "type":"string"
                              }
                           }
                        }
                     }
                  }
               }
            ],
            "responses":{
               "201":{
                  "description":"Succes."
               },
               "400":{
                  "description":"ERROR."
               }
            }
         }
      },
      "/profile/save":{
         "post":{
            "x-swagger-router-controller":"AuthenticationController",
            "operationId":"save",
            "tags":[
               "/profile"
            ],
            "summary":"Save User Profile Details",
            "description":"Save Service Provider Details",
            "parameters":[
               {
                  "name":"user",
                  "in":"body",
                  "description":"Save Service Provider Profile Details.",
                  "schema":{
                     "required":[
                        "user"
                     ],
                     "properties":{
                        "user":{
                           "type":"object",
                           "properties":{
                              "mobileNumber":{
                                 "type":"string"
                              },
                              "firstName":{
                                 "type":"string"
                              },
                              "lastName":{
                                 "type":"string"
                              },
                              "address":{
                                 "type":"string"
                              },
                              "email":{
                                 "type":"string"
                              },
                              "company":{
                                 "type":"string"
                              }
                           }
                        }
                     }
                  }
               }
            ],
            "responses":{
               "201":{
                  "description":"Success."
               },
               "400":{
                  "description":"Error."
               }
            }
         }
      },
      "/profile/update":{
         "put":{
            "x-swagger-router-controller":"AuthenticationController",
            "operationId":"update",
            "tags":[
               "/profile"
            ],
            "summary":"Update User Profile Details",
            "description":"Update Service Provider Details",
            "parameters":[
               {
                  "name":"user",
                  "in":"body",
                  "description":"Save Service Provider Profile Details.",
                  "schema":{
                     "required":[
                        "user"
                     ],
                     "properties":{
                        "user":{
                           "type":"object",
                           "properties":{
                              "mobileNumber":{
                                 "type":"string"
                              },
                              "firstName":{
                                 "type":"string"
                              },
                              "lastName":{
                                 "type":"string"
                              },
                              "address":{
                                 "type":"string"
                              },
                              "email":{
                                 "type":"string"
                              },
                              "company":{
                                 "type":"string"
                              }
                           }
                        }
                     }
                  }
               }
            ],
            "responses":{
               "201":{
                  "description":"Success."
               },
               "400":{
                  "description":"Error."
               }
            }
         }
      },
      "/profile/{mobileNumber}":{
         "get":{
            "x-swagger-router-controller":"AuthenticationController",
            "operationId":"get",
            "tags":[
               "/profile"
            ],
            "summary":"Update User Profile Details",
            "description":"Update Service Provider Details",
            "parameters":[
               {
                  "name":"mobileNumber",
                  "type":"string",
                  "in":"path",
                  "required":true
               }
            ],
            "responses":{
               "200":{
                  "description":"Success.",
                  "schema":{
                     "required":[
                        "user"
                     ],
                     "properties":{
                        "user":{
                           "type":"object",
                           "properties":{
                              "mobileNumber":{
                                 "type":"string"
                              },
                              "firstName":{
                                 "type":"string"
                              },
                              "lastName":{
                                 "type":"string"
                              },
                              "address":{
                                 "type":"string"
                              },
                              "email":{
                                 "type":"string"
                              },
                              "company":{
                                 "type":"string"
                              }
                           }
                        }
                     }
                  }
               },
               "400":{
                  "description":"Error."
               }
            }
         }
      },
      "/skill":{
         "get":{
            "x-swagger-router-controller":"AuthenticationController",
            "operationId":"getSkills",
            "tags":[
               "/skill"
            ],
            "summary":"Get Predefin Skills",
            "description":"Get Predefin Skills",
            "responses":{
               "200":{
                  "description":"Success.",
                  "schema":{
                     "required":[
                        "skills"
                     ],
                     "properties":{
                        "skills":{
                           "type":"array",
                           "items":{
                              "type":"object",
                              "properties":{
                                 "id":{
                                    "type":"string"
                                 },
                                 "title":{
                                    "type":"string"
                                 },
                                 "description":{
                                    "type":"string"
                                 },
                                 "isActive":{
                                    "type":"boolean"
                                 },
                                 "iconUrl":{
                                    "type":"string"
                                 }
                              }
                           }
                        }
                     }
                  }
               },
               "400":{
                  "description":"Error."
               }
            }
         }
      }
   }
}